Set messages correctly in segmented datasets
When updating messages, we need to set the message in the segment it is in already, not the latest segment as was done previously, otherwise we'll duplicate the messages in the dataset.

This update also checks/assigns sequence numbers to messages being updated, and pushes some of the sequence number handling logic in add.py into the firebase_client_wrapper.py, which is possible now due to the updated function in firebase_client_wrapper.py.